% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg19context_freq.R
\docType{data}
\name{hg19context_freq}
\alias{hg19context_freq}
\title{Trinuecleotide frequencies of the human reference genome hg19}
\format{A numeric vector with 32 elements

\describe{
\item{Trinucleotide}{The frequency of a certain trinucleotide context in the 
genome.}}}
\source{
BSgenome.Hsapiens.UCSC.hg19
}
\usage{
hg19context_freq
}
\description{
Trinuecleotide frequencies of the human reference genome hg19
}
\keyword{datasets}
